% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLambda.R
\name{getLambda}
\alias{getLambda}
\title{Get Optimal Lambda value}
\usage{
getLambda(data, lambda = seq(-10, 10, 1/100))
}
\arguments{
\item{data}{a data frame containing the variables.}

\item{lambda}{vector of values of lambda – default (-10, 10) in steps of 0.01.}
}
\value{
a numeric vector
}
\description{
Computes optimal lambda value using \link[MASS]{boxcox} function from the provided data.
}
\examples{
getLambda(mtcars)
}

