% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseDataInfo.R
\name{normaliseDataInfo}
\alias{normaliseDataInfo}
\title{Apply normalisation on a data frame using specific Lambda value}
\usage{
normaliseDataInfo(dat, lambdas)
}
\arguments{
\item{dat}{a data frame containing the variables.}

\item{lambdas}{a numeric vector from the \link[MASS]{boxcox} function}
}
\value{
a list of three elements, the first element, lambdas, is a numeric vector
        containing lambdas, the second element, scaled, is a numeric vector containing
        scale adjustment, the last element, transDat, is a data frame comprising the
        transformed variables.
}
\description{
Apply normalisation on a data frame using specific Lambda value
}
\examples{
\dontrun{normaliseDataInfo(mtcars, getLambda(mtcars, parallel = FALSE))}
normaliseDataInfo(mtcars, getLambda(mtcars, parallel = FALSE))
}
\references{
{
Courtney, M. G. R., and Chang, K. C. (2018) Dealing with non-normality: an introduction and step-by-step guide using R. TEST, doi: 10.1111/test.12154.
}
}
