% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalrShiny.R
\name{normalrShiny}
\alias{normalrShiny}
\title{Shiny application of the normalr}
\usage{
normalrShiny(example = "normalr")
}
\arguments{
\item{example}{name of the shiny apps}
}
\description{
Shiny application of the normalr
}
\examples{
\dontrun{normalrShiny()}
}
