% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{testData}
\alias{testData}
\title{Test dataset for the paper}
\format{
An object of class \code{data.frame} with 957 rows and 9 columns.
}
\usage{
testData
}
\description{
Test dataset for the paper
}
\keyword{datasets}
