% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Apply normalisation on a numeric vector using a specific Lambda value}
\usage{
normalise(x, lambda = 3)
}
\arguments{
\item{x}{a numeric vector to be normalised.}

\item{lambda}{a numeric vector from the \link[MASS]{boxcox} function}
}
\value{
a numeric vector with lambda, scale and the transformed variable
}
\description{
Apply normalisation on a numeric vector using a specific Lambda value
}
\examples{
x <- c(1, 5, 9, 9, 9, 9, 10, 10, 10, 11, 11, 12)
normalise(x, lambda = 3)
}
\references{
{
Courtney, M. G. R., and Chang, K. C. (2018) Dealing with non-normality: an introduction and step-by-step guide using R. TEST, doi: 10.1111/test.12154.
}
}
